requirements-test.txt
pytest>=7.0
PyPDF2>=3.0
Pillow>=9.0